image

+++ Server-side/Product/src/main/java/com/demo/repositories/ImageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.demo.entities.ImageModel;
 
-public interface ImageRepository extends JpaRepository<ImageModel, Long> {
+public interface ImageRepository extends JpaRepository<ImageModel, Integer> {
 
 	Optional<ImageModel> findByName(String name);
